chore: update Go tooling and packages (#11)
Also introduces a workspace file. This is helpful in editors using the 'gopls' language server, as these will now be able to work with multiple Go projects at once.
Use the "tools.go approach" to install the right controller-tools binaries as part of the build process.
By changing "GOBIN" and "PATH" accordingly, we ensure that these binaries are only used locally.