refactor: analytics server/client images to support multiple algorithms (#14)
Changes the correlation server/client code to a more general "analytics" server/client host. In these, multiple methods to aggregate analytics using different algorithms are supported by implementing the respective AnalyticsProvider interfaces and both server and client. This interface allows a provider to define its inputs and output data and allows to configure parameters for the provider. The parameters can then be set as command line parameters when starting a server. Client specific parameters are sent from the server to clients, which also take care of selecting the right provider.

Implementations for federated correlation and histograms are provided.