fix: reconnect on reconcile to handle failures with envoy
In early cluster stage, envoy doesn't yet work as expected and needs a new gRPC
connection once it does. This causes a problem with this controller as it caches
the (falsy) connection once a first one is established. To fix this, we create a
new connection every time the reconcile loop is run.